requirements.txt
flask>=3.0
flask-socketio>=5.3
paho-mqtt>=1.6,<2.0
requests>=2.28
